Blocking *.unequalbrake.com
Unequalbrake *claims* to be a DMCA blocking tool. However, it appears to actually be used to serve adblock detection scripts.

Even were that not the case, Copyright policing is a task for the copyright holders, and not something that should result in javascript being pushed into my browser (why would a copyright violating site willingly serve that JS anyway?)